+++ packages/protos/protos_weebi/CHANGELOG.md
@@ -1,5 +1,15 @@
 # changelog
 
+## 1.3.4 - 2026 may
+
+- simpler offer : entreprise or premium
+
+## 1.3.3 - 2026 may
+
+- add retail total logic in ticket
+- add cel evaluator 
+
+
 ## 1.3.2 - 2026 april
 
 - add business rules in boutique and chain to provide :